+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 
 ### Patch Changes
 
-- Added controlled bee-line measurement props for flat Leaflet maps.
+- Added controlled bee-line measurement props for flat MapLibre maps.
 - Made default heat-map radius and interpolated intensity data-space based.
 
 ## 0.1.3